src/requirements.txt
Flask==2.0.3
ptvsd==4.3.2 # Required for debugging.
google-cloud-bigquery==1.22.0
